Reorganize the code a little bit for readability
